+++ Boussole.Data/BousSOleDbContext.cs
@@ -1,4 +1,5 @@
 using System.Reflection;
+using Boussole.Institutions.Contracts;
 using Boussole.LSO.Contracts.SSO;
 using Boussole.LSO.Contracts.Structure;
 using Microsoft.EntityFrameworkCore;
